Modernize structure of IOMC/ParticleGuns
- Moved producer headers into source file
- Removed use of SealModule.cc
